Add common submodule for shared types used in server api
