+++ .idea/artifacts/relive_simulator_common_js_0_1_1_SNAPSHOT.xml
@@ -1,0 +1,8 @@
+<component name="ArtifactManager">
+  <artifact type="jar" name="relive-simulator-common-js-0.1.1-SNAPSHOT">
+    <output-path>$PROJECT_DIR$/relive-simulator-common/build/libs</output-path>
+    <root id="archive" name="relive-simulator-common-js-0.1.1-SNAPSHOT.jar">
+      <element id="module-output" name="relive-simulator.relive-simulator-common.jsMain" />
+    </root>
+  </artifact>
+</component>
