+++ .idea/artifacts/relive_simulator_core_js_0_1_1_SNAPSHOT.xml
@@ -1,0 +1,6 @@
+<component name="ArtifactManager">
+  <artifact type="jar" name="relive-simulator-core-js-0.1.1-SNAPSHOT">
+    <output-path>$PROJECT_DIR$/relive-simulator-core/build/libs</output-path>
+    <root id="archive" name="relive-simulator-core-js-0.1.1-SNAPSHOT.jar" />
+  </artifact>
+</component>
