+++ .idea/artifacts/relive_simulator_js_0_1_1_SNAPSHOT.xml
@@ -1,0 +1,6 @@
+<component name="ArtifactManager">
+  <artifact type="jar" name="relive-simulator-js-0.1.1-SNAPSHOT">
+    <output-path>$PROJECT_DIR$/relive-simulator-js/build/libs</output-path>
+    <root id="archive" name="relive-simulator-js-0.1.1-SNAPSHOT.jar" />
+  </artifact>
+</component>
