+++ .idea/artifacts/relive_simulator_worker_0_1_1_SNAPSHOT.xml
@@ -1,0 +1,8 @@
+<component name="ArtifactManager">
+  <artifact type="jar" name="relive-simulator-worker-0.1.1-SNAPSHOT">
+    <output-path>$PROJECT_DIR$/relive-simulator-worker/build/libs</output-path>
+    <root id="archive" name="relive-simulator-worker-0.1.1-SNAPSHOT.jar">
+      <element id="module-output" name="relive-simulator.relive-simulator-worker.main" />
+    </root>
+  </artifact>
+</component>
