+++ .idea/modules/relive-simulator-browser/relive-simulator.relive-simulator-browser.test.iml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<module version="4">
+  <component name="TestModuleProperties" production-module="relive-simulator.relive-simulator-browser.main" />
+</module>
