--- relive-simulator-server/src/test/kotlin/server/simulator/relive/qwewqa/xyz/ApplicationTest.kt
@@ -1,30 +1,0 @@
-package xyz.qwewqa.relive.simulator.server
-
-import io.ktor.server.engine.*
-import io.ktor.routing.*
-import io.ktor.http.*
-import io.ktor.http.cio.websocket.*
-import io.ktor.websocket.*
-import java.time.*
-import io.ktor.serialization.*
-import io.ktor.features.*
-import io.ktor.html.*
-import kotlinx.html.*
-import io.ktor.application.*
-import io.ktor.response.*
-import io.ktor.request.*
-import kotlin.test.*
-import io.ktor.server.testing.*
-import xyz.qwewqa.relive.simulator.server.plugins.*
-
-class ApplicationTest {
-    @Test
-    fun testRoot() {
-        withTestApplication({ configureRouting() }) {
-            handleRequest(HttpMethod.Get, "/").apply {
-                assertEquals(HttpStatusCode.OK, response.status())
-                assertEquals("Hello World!", response.content)
-            }
-        }
-    }
-}
