+++ relive-simulator-worker/build.gradle.kts
@@ -1,0 +1,54 @@
+val project_version: String by project
+val ktor_version: String by project
+
+plugins {
+    kotlin("js")
+    id("org.jetbrains.kotlin.plugin.serialization") version "1.5.31"
+}
+
+group = "xyz.qwewqa.relive.simulator"
+version = project_version
+
+repositories {
+    maven { url = uri("https://maven.pkg.jetbrains.space/public/p/kotlinx-html/maven") }
+    mavenCentral()
+    maven { url = uri("https://maven.pkg.jetbrains.space/public/p/ktor/eap") }
+}
+
+dependencies {
+    implementation(kotlin("stdlib-js"))
+    implementation(project(":relive-simulator-core"))
+}
+
+kotlin {
+    js(IR) {
+        browser {
+            webpackTask {
+                cssSupport.enabled = true
+            }
+
+            runTask {
+                cssSupport.enabled = true
+            }
+
+            testTask {
+                useKarma {
+                    useChromeHeadless()
+                    webpackConfig.cssSupport.enabled = true
+                }
+            }
+        }
+        compilations.all {
+            kotlinOptions {
+                freeCompilerArgs += "-Xir-property-lazy-initialization"
+            }
+        }
+        binaries.executable()
+    }
+}
+
+tasks.withType(org.jetbrains.kotlin.gradle.dsl.KotlinCompile::class) {
+    kotlinOptions.run {
+        freeCompilerArgs = freeCompilerArgs + "-Xopt-in=kotlin.RequiresOptIn"
+    }
+}
