+++ relive-simulator-worker/src/main/kotlin/xyz/qwewqa/relive/simulator/js/Main.kt
@@ -1,0 +1,12 @@
+package xyz.qwewqa.relive.simulator.js
+
+import org.w3c.dom.DedicatedWorkerGlobalScope
+import org.w3c.dom.MessageEvent
+
+external val self: DedicatedWorkerGlobalScope
+
+fun main() {
+    self.onmessage = { ev: MessageEvent ->
+
+    }
+}
