+++ settings.gradle.kts
@@ -4,3 +4,5 @@ include("relive-simulator-core")
 include("relive-simulator-server")
 include("relive-simulator-client")
 include("relive-simulator-browser")
+include("relive-simulator-worker")
+include("relive-simulator-common")
